Better AppleScriptValue instances
Previously we were using "show" in instances for String, Text, and
Strict.Text. However, this causes failures in most values spliced in
between quotes. Therefore, we no longer "show" the textual values and
instead simply use them as is.

If quotes around the string are desired it must be shown manually or
added manually.